(CONT-413) Bump additional ruby version
Ruby version 2.7.6 had a CVE reported last year. As a result references
in the puppet-runtime project were bumped to 2.7.7.

For this compoent we hard code the additional_rubies hash key. Prior to
this change it was set to 2.7.6 which was causing build failures.

This change updates the key to 2.7.7 so that the look up will work as
expected.