(CONT-734) Re-enable rubyopt patch
There is a bug in bundler where it will fail with
`invalid switch in RUBYOPT: -F (RuntimeError)` if a given path contained
spaced.. This is common on Windows.

Prior to this change it had been disabled as it appeared to be
resolved.. however after an upgrade of bundler in pdk-vanagon it has
re-appeared.

This change re-enables the patch.